major refactoring, StructuralSymmetry->ReferenceStructure
