+++ ramannoodle/spectrum/spectrum_utils.py
@@ -3,71 +3,10 @@
 import numpy as np
 from numpy.typing import NDArray
 
-from ..globals import BOLTZMANN_CONSTANT
-from ..exceptions import verify_ndarray_shape, verify_ndarray, get_type_error
+from ramannoodle.exceptions import verify_ndarray_shape, verify_ndarray, get_type_error
 
 
-def get_bose_einstein_correction(
-    wavenumbers: NDArray[np.float64], temperature: float
-) -> NDArray[np.float64]:
-    """Calculate Bose-Einstein spectral correction.
-
-    Parameters
-    ----------
-    wavenumbers
-    temperature
-        in kelvin
-
-    Returns
-    -------
-    :
-        Correction factor for each wavenumber.
-
-    Raises
-    ------
-    ValueError
-
-    """
-    try:
-        if temperature <= 0:
-            raise ValueError(f"invalid temperature: {temperature} <= 0")
-    except TypeError as exc:
-        raise get_type_error("temperature", temperature, "float") from exc
-    try:
-        energy = wavenumbers * 29979245800.0 * 4.1357e-15  # in eV
-        return 1 / (1 - np.exp(-energy / (BOLTZMANN_CONSTANT * temperature)))
-    except TypeError as exc:
-        raise get_type_error("wavenumbers", wavenumbers, "ndarray") from exc
-
-
-def get_laser_correction(
-    wavenumbers: NDArray[np.float64], laser_wavenumber: float
-) -> NDArray[np.float64]:
-    """Calculate conventional laser-wavenumber-dependent spectral correction.
-
-    Parameters
-    ----------
-    wavenumbers
-    laser_wavenumber
-
-    Returns
-    -------
-    :
-        Correction factor for each wavenumber.
-
-    """
-    try:
-        if laser_wavenumber <= 0:
-            raise ValueError(f"invalid laser_wavenumber: {laser_wavenumber} <= 0")
-    except TypeError as exc:
-        raise get_type_error("laser_wavenumber", laser_wavenumber, "float") from exc
-    try:
-        return ((wavenumbers - laser_wavenumber) / 10000) ** 4 / wavenumbers
-    except TypeError as exc:
-        raise get_type_error("wavenumbers", wavenumbers, "ndarray") from exc
-
-
-def convolve_intensities(
+def convolve_spectrum(
     wavenumbers: NDArray[np.float64],
     intensities: NDArray[np.float64],
     function: str = "gaussian",
